backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-multipart>=0.0.6
SpeechRecognition>=3.10.0
pydub>=0.25.1
